refactoring createDocument
