+++ efcms-service/src/documents/services/documentBlobDAO.js
@@ -1,6 +1,20 @@
 const { create: getPersistence } = require('../../persistence/entityPersistenceFactory');
 
 exports.getDocumentDownloadUrl = async ({ documentId }) => {
-  return await getPersistence('files')
+  return getPersistence('files')
     .getDocumentDownloadUrl({ documentId });
 }
+
+exports.createUploadPolicy = async () => {
+  return getPersistence('files')
+    .createUploadPolicy();
+};
+
+exports.createDocument = async ({ userId, documentType }) => {
+  if (!userId || !documentType) {
+    throw new Error('documentType and userId are required');
+  }
+
+  return getPersistence('documents')
+    .createDocument({ userId, documentType });
+};
