+++ efcms-service/src/persistence/aws/awsDynamoPersistence.js
@@ -1,3 +1,5 @@
+const uuidv4 = require('uuid/v4');
+const environment = require('../../environment');
 const { get: getEnv } = require('../../environment');
 const client = require('../../middleware/dynamodbClientService');
 
@@ -12,4 +14,23 @@ exports.updateCase = async ({ caseToUpdate }) => {
   };
   return client.put(params);
 }
+
+exports.createDocument = ({ userId, documentType }) => {
+  const documentId = uuidv4();
+  const params = {
+    TableName: environment.get('DOCUMENTS_TABLE'),
+    Item: {
+      documentId: documentId,
+      createdAt: new Date(),
+      userId: userId,
+      documentType: documentType,
+    },
+    ConditionExpression: 'attribute_not_exists(#documentId)',
+    ExpressionAttributeNames: {
+      '#documentId': 'documentId',
+    },
+  };
+
+  return client.put(params);
+};
